Avoid using Arc::make_mut for FrameMEStats
This change comes from an attempt to fix a new
clippy lint from Rust 1.48. The simple fix would have
involved changing Arc::make_mut to Arc::get_mut.
However, that caused panics in several tests,
which means that Arc::make_mut was not actually
mutating the original data in many cases.
So, this refactor avoids a mutable reference
to the FrameMEStats entirely.

In order to satisfy lifetime constraints,
the best and most efficient way to handle this
was to make a variant of MotionVector which
uses Atomic integers to store its data.
This allows its values to be changed safely
across threads without requiring a mutable reference
or a Mutex.

BD Rate seems to be unaffected, although this has
approximately a 1% speedup at speed 5, most likely
due to the avoidance of cloning the interior of the Arc
when make_mut failed to create a mutable reference.
At speed 10 the overall encode speed appears unaffected.