Remove unnecessary min calls in nz functions
The nz functions force a max level of 3 for each of the level inputs,
but currently we accomplish this by initializing the levels to a max of
127, then capping them to 3 when we read them.

Since we do not read from these levels in any place other than this
function, we can simply initialize them with a max of 3, and avoid the
comparisons in `get_nz_mag`.

Since this is a hot function, this reduces total encoding time by
about 2% at speed 2, based on benchmarking a small set of clips from
objective-1-fast.